+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog of threedi-scenario-downloader
 ----------------
 
 - Added support for sub-endpoints "/api/v4/scenarios/{uuid}/results/damage/" and "/api/v4/scenarios/{uuid}/results/arrival/"
+- Now added the use of keyword arguments for "resolution", "projection","bbox" and "time".
+- Bugfix with width and height
 
 
 1.0 (2023-05-15)